Rename v2 event api files
