+++ ui/src/api/events/query/EventStreamingSubscriber.tsx
