+++ ui/src/api/events/query/stream.ts
