+++ ui/src/api/events/schema/filters.ts
