+++ ui/src/api/events/schema/index.ts
