+++ ui/src/pages/namespace/Events/History/SendEvent.tsx
@@ -7,7 +7,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "~/design/Dialog";
-import { NewEventSchema, NewEventSchemaType } from "~/api/eventsv2/schema";
+import { NewEventSchema, NewEventSchemaType } from "~/api/events/schema";
 import { PlusCircle, Radio } from "lucide-react";
 import { SubmitHandler, useForm } from "react-hook-form";
 
@@ -15,7 +15,7 @@ import Button from "~/design/Button";
 import { Card } from "~/design/Card";
 import Editor from "~/design/Editor";
 import { exampleEvent } from "./exampleEvent";
-import { useSendEvent } from "~/api/eventsv2/mutate/sendEvent";
+import { useSendEvent } from "~/api/events/mutate/sendEvent";
 import { useState } from "react";
 import { useTheme } from "~/util/store/theme";
 import { useTranslation } from "react-i18next";
